+++ src/services/mod.rs
@@ -1,4 +1,5 @@
 pub mod db;
+pub mod io;
 pub mod server;
 pub mod utils;
 pub mod watcher;